view matches and refactoring game layout
