+++ react/tic-tac-toe/src/app/screens/Home/components/Routes/index.js
@@ -3,13 +3,13 @@ import { Switch, Route } from 'react-router-dom';
 
 import Game from '~screens/Home/screens/Game';
 
-import Matchesitos from '~screens/Home/screens/Matches';
+import Matches from '~screens/Home/screens/Matches';
 
 function Routes() {
   return (
     <Switch>
       <Route path="/game" exact component={Game} />
-      <Route path="/matches" exact component={Matchesitos} />
+      <Route path="/matches" exact component={Matches} />
     </Switch>
   );
 }
