--- react/tic-tac-toe/src/app/screens/Home/screens/Game/components/Matches/styles.module.scss
@@ -1,11 +1,0 @@
-@import 'src/scss/colors.scss';
-$square-border: $dust-gray;
-$text-player: $black;
-.match {
-  border: 1px solid $square-border;
-  border-radius: 5px;
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 10px;
-  padding: 1rem;
-}
