+++ react/tic-tac-toe/src/app/screens/Home/screens/Game/components/Square/index.js
@@ -1,5 +1,6 @@
 import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
+import cn from 'classnames';
 
 import styles from './styles.module.scss';
 
@@ -8,7 +9,11 @@ function Square({ onClick, position, value }) {
     onClick(position);
   }, [position]);
   return (
-    <button type="button" className={styles.square} onClick={handleClick}>
+    <button
+      type="button"
+      className={cn(styles.square, { [styles.nextPlayer]: value === 'X' })}
+      onClick={handleClick}
+    >
       {value}
     </button>
   );
