+++ react/tic-tac-toe/src/app/screens/Home/screens/Game/index.js
@@ -5,7 +5,6 @@ import { calculateWinner } from '~utils/game';
 import styles from './styles.module.scss';
 import Board from './components/Board';
 import Moves from './components/Moves';
-import Matches from './components/Matches';
 
 class Game extends Component {
   state = {
@@ -56,10 +55,6 @@ class Game extends Component {
         <div className={styles.gameInfo}>
           <div>{status}</div>
           <ol>{history.map(this.renderMoves)}</ol>
-        </div>
-        <div className={styles.gameInfo}>
-          <div>Matches</div>
-          <Matches />
         </div>
       </div>
     );
