+++ react/tic-tac-toe/src/app/screens/Home/screens/Matches/index.js
@@ -1,24 +1,43 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { useEffect } from 'react';
+import Spinner from 'react-spinkit';
+import { useSelector, useDispatch } from 'react-redux';
 
-function Matchesitos(props) {
+import gameActions from '~redux/game/actions';
+
+import styles from './styles.module.scss';
+
+function Matches() {
+  const dispatch = useDispatch();
+  const { loading, matches } = useSelector(state => state.game);
+
+  useEffect(() => {
+    dispatch(gameActions.getMatches());
+  }, []);
+
   return (
-    <div>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
-      <h1>matcheeesss</h1>
+    <div className={styles.container}>
+      {loading ? (
+        <Spinner name="double-bounce" />
+      ) : (
+        <ol className={styles.matches}>
+          {matches &&
+            matches.map(match => (
+              <li className={styles.match} key={match.createdAt}>
+                <h3 className={styles.player}>
+                  Jugador uno: <span>{match.player_one}</span>
+                </h3>
+                <h3 className={styles.player}>
+                  Jugador dos: <span>{match.player_two}</span>
+                </h3>
+                <h3 className={styles.winner}>
+                  Ganador: <span>{match.winner}</span>
+                </h3>
+              </li>
+            ))}
+        </ol>
+      )}
     </div>
   );
 }
 
-Matchesitos.propTypes = {};
-
-export default Matchesitos;
+export default Matches;
