+++ react/tic-tac-toe/src/app/screens/Home/screens/Matches/styles.module.scss
@@ -1,0 +1,31 @@
+@import 'src/scss/colors.scss';
+$text-player: $black;
+$text-winner-color: $wolox-green;
+.match {
+  border-radius: 5px;
+  box-shadow: rgba(0, 0, 0, 0.1) 0 4px 12px;
+  display: flex;
+  flex-basis: 200px;
+  flex-direction: column;
+  margin-bottom: 20px;
+  padding: 16px;
+}
+
+.container {
+  padding: 32px;
+}
+
+.matches {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+}
+
+.player {
+  color: $text-player;
+  margin-bottom: 10px;
+}
+
+.winner {
+  color: $text-winner-color;
+}
