+++ react/tic-tac-toe/src/scss/_colors.scss
@@ -2,6 +2,7 @@ $white: #FFF;
 $dust-gray: #999;
 $black: #000;
 $wolox-blue: #00A7E8;
+$wolox-green: #9FC739;
 $gainsboro-gray: #DDD;
 $silver-gray: #CCC;
 $red: #86181D;
